further solving optimizations
Benchmark is down from 28 seconds to 8 seconds.

Now, we don't incur any of the cost that the map[State]bool was giving us. Also, allocating stuff for the list of previous moves was taking up some time, so using a linked-list gained some performance.

The solver also uses much less memory than it did before (300MB now, as compared to a GB or two before).