neural network actually works well!
By using sinusoidal activations, we can capture the periodic behavior of the integers mod 2. The results are astounding.

The network outputs a distribution of solutions. On average, the correct solution has around a 35% probability, so sampling the output of the network many times will almost surely yield a solution in a short period of time.